Parse track names from tags in SongLoader
