+++ src/main/java/com/github/nianna/karedi/txt/loader/SongLoader.java
@@ -1,10 +1,12 @@
 package com.github.nianna.karedi.txt.loader;
 
 import com.github.nianna.karedi.I18N;
+import com.github.nianna.karedi.song.Song;
+import com.github.nianna.karedi.song.tag.Tag;
 import com.github.nianna.karedi.txt.parser.Parser;
 import com.github.nianna.karedi.txt.parser.element.InvalidSongElementException;
 import com.github.nianna.karedi.txt.parser.element.VisitableSongElement;
-import com.github.nianna.karedi.song.Song;
+import com.github.nianna.karedi.util.MultiplayerTags;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -38,7 +40,9 @@ class SongLoader {
 				.filter(line -> !line.isBlank())
 				.map(this::parseLine)
 				.forEach(songBuilder::buildPart);
-		return songBuilder.getResult();
+		Song song = songBuilder.getResult();
+		extractTrackNamesFromTags(song);
+		return song;
 	}
 
 	private VisitableSongElement parseLine(String line) {
@@ -50,4 +54,17 @@ class SongLoader {
 		}
 	}
 
+	private void extractTrackNamesFromTags(Song song) {
+		List<Tag> trackNameTags = song.getTags().stream()
+				.filter(MultiplayerTags::isANameTag)
+				.toList();
+		trackNameTags.forEach(tag -> renameTrack(song, tag));
+	}
+
+	private static void renameTrack(Song song, Tag tag) {
+		MultiplayerTags.getTrackNumber(tag).ifPresent(number -> {
+			song.removeTag(tag);
+			song.renameTrack(number, tag.getValue());
+		});
+	}
 }
