+++ src/test/java/com/github/nianna/karedi/context/SongNormalizerTest.java
@@ -1,7 +1,6 @@
 package com.github.nianna.karedi.context;
 
 import com.github.nianna.karedi.song.Song;
-import com.github.nianna.karedi.song.SongTrack;
 import com.github.nianna.karedi.song.tag.Tag;
 import com.github.nianna.karedi.song.tag.TagKey;
 import org.junit.Assert;
@@ -59,21 +58,4 @@ public class SongNormalizerTest {
         Assert.assertEquals(1, song.getTrack(0).getPlayer().intValue());
     }
 
-    @Test
-    public void shouldSetTrackNamesFromTags() {
-        Song song = new Song();
-        SongTrack firstTrack = new SongTrack(1);
-        SongTrack secondTrack = new SongTrack(2);
-        song.addTrack(firstTrack);
-        song.addTrack(secondTrack);
-
-        song.addTag(new Tag("DUETSINGERP1", "Elvis Presley"));
-        song.addTag(new Tag("DUETSINGERP2", "Freddie Mercury"));
-        song.addTag(new Tag("DUETSINGERP3", "Ignored"));
-
-        songNormalizer.normalize(song);
-
-        Assert.assertEquals("Elvis Presley", firstTrack.getName());
-        Assert.assertEquals("Freddie Mercury", secondTrack.getName());
-    }
 }
